pyproject.toml
[project]
name = "spy-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.11"
dependencies = [
    "macholib>=1.16.3",
    "modulegraph>=0.19.6",
    "numpy>=2.3.2",
    "pip>=25.2",
    "setuptools>=80.9.0",
    "sounddevice>=0.5.2",
    "Pillow>=10.0.0",
    "scipy>=1.12.0",
]
